MemoryView: replace slots with functions and rename them
The slots are connected only via the function pointer-based syntax, so
they can be ordinary member functions. Remove the no longer accurate
word "slot" from their names.

Replace the trivial, used-once slotStateChanged() with a lambda.

Make the public function MemoryView::debuggerStateChanged() private to
reuse the applicable-to-it documentation of the removed
slotStateChanged() and because only MemoryView itself uses
debuggerStateChanged().