Remove useless MemoryView::isOk() that always returns true
f5b33ab4330944b7fa38714e9410427fc2fcfdcb replaced the setup of
khexedit2_widget that set it to nullptr on failure with the following
assignment:
    m_memViewView = new Okteta::ByteArrayColumnView(this);
and started returning m_memViewView instead of khexedit2_widget from
MemoryView::isOk(). So now isOk() is called only after a non-null value
is assigned to m_memViewView, and therefore it always returns true.

In addition to removing the useless function isOk(), do not check
m_memViewView and (similarly initialized to a new ByteArrayModel in
initWidget()) m_memViewModel for null needlessly. And remove a
potentially misleading initialization of m_memViewView to nullptr.