MemoryView: track debugger state more reliably
Multiple issues make MemoryView's tracking of the debugger state
unreliable:
1. MemoryView() initializes m_debuggerState to 0, and therefore assumes
   that a debuggee is running when it is created. However, GDB tool
   views, including the Memory tool view, are created in the constructor
   of GDB::DebugSession when neither the debugger nor a debuggee has
   started.
2. MemoryView() does not connect to the signals of a current debug
   session, and therefore relies on being created before a debug session
   becomes current.
3. MemoryView::currentSessionChanged() fails to update m_debuggerState
   when the current session becomes nullptr, and therefore relies on the
   debugger state to become not-running at the same time or earlier.
4. MemoryView::currentSessionChanged() does not update m_debuggerState
   when the current session becomes non-null either, and therefore
   relies on the newly current session to change its debugger state soon
   after.

I do not know how to reproduce incorrect behavior due to these issues.
But the unnecessary assumptions make MemoryView fragile in the face of
changes to the time of construction and the reuse of debugger plugin
tool views, as well as disconnection from the signals of a no longer
current debug session, which I plan to implement.

Get rid of the issues by updating the stored debugger state more often.

Store `bool m_appHasStarted` instead of `int m_debuggerState` in
MemoryView to simplify the code and to avoid having to set
m_debuggerState to a bogus value s_appNotStarted when there is no
current debug session.