Debugger output widgets: disable UI when no current session
GDBOutputWidget::slotStateChanged() disables the Interrupt button and
the command editor line if the debugger is not running. Disable the same
two widgets when there is no current debug session instead of waiting
for the no longer current session to change its state. The disabling
will become more important when a subsequent commit disconnects
GDBOutputWidget from no longer current sessions.

Extract handleDebuggerNotRunning() and remove the unused first parameter
from slotStateChanged() to avoid passing bogus debugger state arguments
s_none and s_dbgNotStarted.

Remove the no longer needed call
slotStateChanged(s_none, s_dbgNotStarted) from GDBOutputWidget() because
the same constructor calls GDBOutputWidget::currentSessionChanged(),
which now always updates the UI according to the current debugger state.

Rename slotStateChanged() to debuggerStateChanged() for consistency with
MemoryView. Make debuggerStateChanged() a private function instead of a
public slot because it is used only by GDBOutputWidget itself and is
connected only via the function pointer-based syntax. Rename the
parameter of debuggerStateChanged() from newStatus to newState for
consistency with the name of the corresponding parameter of the signal,
to which this function is connected.

The implementation of DebuggerConsoleView was copied from
GDBOutputWidget and adjusted. The relevant code is practically the same
in the two widgets, and this commit made analogous changes to it. So the
explanation above applies to the DebuggerConsoleView changes as well.