MemoryView: don't create a QMenu on the stack
A QMenu object created on the stack in MemoryView::contextMenuEvent()
has a parent (this), and exec() is called on it. So the menu object can
be destroyed before exec() returns. QMenuPrivate::exec() even guards the
QMenu object with a QPointer and returns nullptr if the guard is null
after the nested event loop returns. When the destructor of the already
destroyed QMenu object is inevitably invoked at the end of its scope,
the second destruction causes undefined behavior and likely a crash.

I have managed to reproduce such a crash with the following steps:
1. Prepare a debuggee that calls QThread::sleep(5) then exits.
2. Start debugging the debuggee with GDB; show the Memory tool view and
   right-click a Memory View widget while the debuggee is still running.
3. Once the 5 seconds elapse, the debuggee exits, the debug session
   finishes and is destroyed. The destructor of GDB::DebugSession
   invokes CppDebuggerPlugin::unloadToolViews(), which unregisters the
   Memory tool view and destroys its tool view widget. This invokes the
   destruction of the Memory View widget, which destroys its child
   widget QMenu. All this occurs within the nested event loop of the
   QMenu::exec() call, hence KDevelop crashes with the following output
   (curiously, the crash occurs during the first destruction of the
   QMenu - probably because GCC notices that an object created on the
   stack is destroyed elsewhere, which invokes undefined behavior):
    double free or corruption (out)
    *** Program received signal SIGABRT (Aborted) ***

Create the QMenu object on the heap and set the attribute
Qt::WA_DeleteOnClose on it to prevent this crash and to not leak the
menu object either.