Return nullptr instead of empty menu from createStandardContextMenu()
Returning a newly constructed empty menu if there are no actions to show
from StandardDocumentationView::createStandardContextMenu(), then
destroying it in StandardDocumentationView::contextMenuEvent(), is
wasteful.

QtHelpDocumentation::viewContextMenuRequested() does not check whether
the returned menu is empty and shows it in any case. Returning nullptr
instead of an empty menu reveals and forces avoiding this useless waste
of resources (an empty context menu does not actually appear, but
QMenuPrivate::popup() detects the emptiness and returns only after quite
some work).

Returning nullptr is consistent with the behavior of
QWidgetTextControl::createStandardContextMenu(), to which the same-name
member functions createStandardContextMenu() of QTextEdit and
QPlainTextEdit delegate.