Destroy QMenu via WA_DeleteOnClose rather than QScopedPointer
A QMenu object can be destroyed before a call to exec() on it returns.
QMenuPrivate::exec() even guards the QMenu object with a QPointer and
returns nullptr if the guard is null after the nested event loop
returns. When QScopedPointer<QMenu> invokes the destructor of the
already destroyed QMenu object at the end of its scope, the second
destruction causes undefined behavior and likely a crash.

Replace all such unsafe usages of QScopedPointer<QMenu> (except for
RuntimeController::m_runtimesMenu, which is not function-scoped) with
raw QMenu pointers, and set the attribute Qt::WA_DeleteOnClose before
calling exec() on them. There are no early returns before the menu is
shown in any of the modified functions, hence none of the menu objects
is leaked.