Check createStandardContextMenu() for null and add separator
QWidgetTextControl::createStandardContextMenu(), to which the same-name
member functions createStandardContextMenu() of QTextEdit and
QPlainTextEdit delegate, returns nullptr if there are no actions to
show. DebuggerConsoleView::showContextMenu() and
GDBOutputWidget::gdbViewContextMenuRequested() call
createStandardContextMenu() on a QTextEdit and a QPlainTextEdit
respectively, but unconditionally dereference the returned QMenu pointer
instead of checking it for null.

GDBOutputWidget::gdbViewContextMenuRequested() and
ProjectItemLineEdit::showCtxMenu() do not insert a separator between
standard and custom context menu actions.

Use KDevelop::prepareStandardContextMenuToAddingCustomActions() in the 3
aforementioned places to automatically check for null and consistently
insert a separator before custom actions. The check is not necessary in
showCtxMenu(), because the current implementation of
QLineEdit::createStandardContextMenu() never returns nullptr, but it
does not hurt either and increases robustness in the face of Qt changes.