Debugger output widgets: pass position to createStandardContextMenu()
GDBOutputWidget calls the overload of
QPlainTextEdit::createStandardContextMenu() with zero arguments, the
documentation for which recommends using the overload that takes a
single position argument instead.

A FIXME comment in DebuggerConsoleView::showContextMenu() questions the
coordinates of the position argument it passes to
QTextEdit::createStandardContextMenu(). Q(Plain|)TextEdit::document()
return a QTextDocument*, which inherits QObject and consequently does
not offer the mapFrom* member functions of QWidget. So it is unclear
what the "document coordinates" in the documentations of
Q(Plain|)TextEdit::createStandardContextMenu() refer to and how to
convert a position in the viewport coordinates provided by the signal
QWidget::customContextMenuRequested() to them.

I have tested adding Qt::LinksAccessibleByMouse to the text interaction
flags and inserting HTML links (bypassing the calls to
QString::toHtmlEscaped()) into the Q(Plain|)TextEdit among the real
HTML-escaped ouptut of GDB/MI or LLDB-MI. Then right-clicking the links
makes the context menu action Copy Link Location enabled, while
right-clicking outside of the links - disabled. I have also tested
horizontal coordinate offsets by setting the word wrap mode of
Q(Plain|)TextEdit to QTextOption::WordWrap and setting their layout
direction to Qt::RightToLeft.

The testing and analysis of Qt source code revealed the inconvenience of
requiring user code to convert to the document coordinates and a bug in
the relevant implementation of QPlainTextEdit. Reported these issues in
QTBUG-138099. This commit passes the positions in the coordinates that
work correctly with the latest Qt version 6.9.1.

Rename DebuggerConsoleView::showContextMenu() to
textViewContextMenuRequested() to document why this slot customizes the
standard context menu of m_textView. Rename the argument of this slot
from pos to viewportPosition to document the coordinates of the position
provided by the signal QTextEdit::customContextMenuRequested(). Convert
this protected slot into a private member function because it is
connected only via the function pointer-based syntax and is used only by
DebuggerConsoleView itself.