popup() rather than exec() several context menus
Each of the 5 functions modified in this commit either handles a context
menu event or is invoked when the signal
QWidget::customContextMenuRequested() is emitted. All of these functions
call QMenu::exec() in the end and do not use its return value. Call
QMenu::popup() instead, which is safe in such cases, to avoid a
potentially problematic nested event loop spawned by exec().