Debugger output widgets: clear pending output too
The member function clear() of GDBOutputWidget and DebuggerConsoleView
is invoked when the debugger of a finishing debug session is stopped.
This function clears all output of the no longer relevant session.
Except for the output added in MIDebugSession::stopDebugger() itself
(which contains at least the gdb-exit command) because
GDBOutputWidget::showLine() and DebuggerConsoleView::appendLine() append
the received output asynchronously to improve performance.

As a result, starting a new GDB or LLDB debug session while the previous
session of the same type (GDB or LLDB respectively) is running and while
the Show Internal Commands checkbox is checked, leaves extra output like
the following at the top of the GDB tool view:
    (gdb) 28-gdb-exit
    (gdb) quit
and the LLDB Console tool view:
    (lldb) 21-gdb-exit
    (lldb) quit
Toggling the Show Internal Commands checkbox twice gets rid of this
output left over from the previous debug session because the function
clear() has removed it from GDBOutputWidget::m_allCommands or
DebuggerConsoleView::m_allOutput.

Clear the pending output and stop the flush-pending timer in the member
function clear() of GDBOutputWidget and DebuggerConsoleView to
completely remove all output of a finishing debug session.