GDBOutputWidget: don't store unformatted command output
The "raw" command lists and their documentation probably predate even
Qt 4 because QTextEdit::toPlainText() is available in Qt 4.0 and
QPlainTextEdit::toPlainText() is available in Qt 4.4 (when the class
QPlainTextEdit was introduced).

This removal of an obsolete workaround also fixes the following bug: if
the user starts a new GDB session while the previous GDB session is
running, the Copy All action copies the output of both the previous and
the current debug session. That's because GDBOutputWidget::clear() fails
to clear m_userCommandsRaw and m_allCommandsRaw.

Use QPlainTextEdit::toPlainText() instead of the removed data members.
Now the Copy All action is guaranteed to be an equivalent shortcut for
triple-clicking to select all text and triggering the Copy action. Note
that unlike the triple-clicking and the Copy All action, the sequence of
two actions - Select All and Copy - does not affect the primary
selection. In any case, the contents is conveniently copied as plain
text, without escape sequences.