GDBOutputWidget: color stderr and log messages as warnings
GDBOutputWidget connects GDBOutputWidget::slotInternalCommandStdout()
instead of the unused GDBOutputWidget::slotReceivedStderr() to the
signal DebugSession::debuggerInternalOutput(). The definition of
slotReceivedStderr() is analogous to that of
DebuggerConsoleView::receivedStderr(), which is connected to the signal
debuggerInternalOutput(). But when I tried connecting
slotReceivedStderr() to debuggerInternalOutput(), the user command view
became cluttered with regular non-error GDB/MI log messages colored as
errors (red) - see the examples in the added comment.

So keep showing the stderr and log GDB messages only among internal
command output. But color such messages as warnings (brown) so as to
distinguish them from regular output without implying that they are
errors.