DisassembleWidget: don't deactivate when debugger stops
Starting a new debug session while the previous one is running and while
the Disassemble/Registers tool view is visible prevents all further
updates of the tool view until it is hidden and shown again.

Do not deactivate the Disassemble/Registers tool view when the debugger
stops and consequently the signal MIDebuggerPlugin::reset() is emitted
in order to eliminate this bug. The deactivation is pointless anyway
because when the debugger is stopped, its session does not emit the
signals showStepInSource() and showStepInDisassemble() that an inactive
Disassemble/Registers tool view ignores.