DisassembleWidget: always disassemble on flavor change
The Disassemble/Registers tool view is almost certainly visible when
DisassembleWidget::setDisassemblyFlavorHandler() is invoked because the
invocation of the handler is an eventual outcome of triggering a
disassembly flavor action in the context menu of the disassembled memory
region view. The DisassembleWidget is active if and only if the tool
view is visible.

If the Disassemble/Registers tool view happens to become hidden while a
disassembly flavor change is being applied, skipping disassembling a
memory region results in a stale view with an outdated flavor in case
the tool view becomes visible again before the current address changes.
That's because DisassembleWidget::slotActivate() checks whether the
current address is among the displayed addresses and skips
re-disassembling if so.

An alternative fix is to set DisassembleWidget::address_ to an invalid
value, e.g. 0, if DisassembleWidget::active_ is false in
setDisassemblyFlavorHandler() so as to ensure that the eventual call to
DisassembleWidget::slotActivate() re-disassembles. But complicating the
code to slightly improve performance in this unlikely scenario is not
worthwhile.