Accept unsigned long as valid address in SelectAddressDialog
DisassembleWidget always converts an address string to unsigned long.
But SelectAddressDialog, which is used only by DisassembleWidget to ask
the user for an address, accepts any long long instead. Consequently, a
negative number is accepted and converted to its two's complement. And
a valid address greater or equal to 2^63 is not accepted.

Extract two helper functions and use them consistently to unify
conversion from string to unsigned long address, and thus fix the bug as
well as prevent such mistakes in the future.

Instead of ignoring address conversion errors, assert that the
conversions succeed. This way, if a conversion unexpectedly fails, we
shall learn about it and handle the error properly.