Let SelectAddressDialog manage its OK button itself
The enabled state of the OK button is updated whenever the selected
address changes. No need to also update it each time the dialog is
shown. So disable the button in the constructor instead.

Make public member functions hasValidAddress() and updateOkState() that
are now used only by SelectAddressDialog itself private.