DisassembleWidget: allow disassembling around any address
Attempting to disassemble around (after) one of the currently displayed
addresses within the disassembled memory region keeps the region intact
and selects the row of the address of the currently executed instruction
(DisassembleWidget::address_). This behavior is surprising and not
useful. Remove the address checks in order to allow the user to shift
the displayed disassembled memory region forward by any number of
instructions.