DisassembleWidget: always display correct memory region
Bugs fixed by this commit:
1. Hiding the Disassemble/Registers tool view while a memory region is
   disassembled in it, letting the debuggee continue and pause at a
   different location, then showing the tool view again keeps displaying
   the old disassembled memory region and the stale current address.
   This occurs because while the Disassemble/Registers tool view is
   hidden, DisassembleWidget::active_ is false, in which case
   DisassembleWidget::update() returns early without updating the value
   of DisassembleWidget::address_; showing the tool view invokes
   DisassembleWidget::slotActivate(), which does not disassemble a new
   memory region if address_ is inside the currently displayed region.
2. Triggering the Change Address action in the context menu of the
   disassembled memory region view and selecting a starting address such
   that the current address is not within the new 256-byte memory
   region, then hiding and showing the Disassemble/Registers tool view
   replaces the manually configured memory region with another region
   that starts at the current address. That's because slotActivate()
   always ensures that the current address is within the displayed
   region, even if nothing changes while the tool view is hidden.
3. Selecting a disassembly flavor in the context menu of the
   disassembled memory region view disassembles a memory region that
   starts at the current address instead of keeping the currently
   displayed region and only updating its disassembly flavor.

Always update the current address in DisassembleWidget::update(), even
if the tool view is hidden. Converting a string to an integer is not
prohibitively expensive.

While the tool view is hidden and active_ is false, only one function is
inhibited - DisassembleWidget::update(). So when the tool view becomes
visible and DisassembleWidget::slotActivate() is invoked, it should
perform the work of DisassembleWidget::update() only if the latter has
been invoked at least once while inhibited. Introduce and use bool
DisassembleWidget::m_upToDate to keep track of such invocations.

Introduce a helper class StoredAddress to store string representations
of addresses in addition to the integer representations.

Use the string representations of the bounding addresses to
re-disassemble the currently displayed region when the user selects a
disassembly flavor. If no region is currently displayed, do not
re-disassemble anything, because the updated flavor will be used
automatically in the next disassembly.

Use the string representation of the current address instead of
executing an extra GDB command to find it each time. This optimization
makes it possible to require the `from` argument of
DisassembleWidget::disassembleMemoryRegion() to be a valid address and
to remove the no longer needed member function
DisassembleWidget::updateExecutionAddressHandler().

Clear out all addresses of the previous debug session when the current
session changes. Otherwise, the increased use of stored addresses in
this commit would introduce multiple regressions because of stale
addresses. A (not unreasonable) conspicuous side effect of this change
is that starting a new debug session while the previous one is running
immediately clears the displayed disassembled memory region instead of
keeping it until the new session enters the paused state.

Improve the names of several data members and of a local variable (cmd).

Use QLatin1String instead of QStringLiteral in disassembleMemoryRegion()
to optimize.