DisassembleWidget: don't revert the range of displayed region
If the user selects a disassembly flavor in the context menu of the
disassembled memory region view while the current debug session is in
the active state (i.e. the debuggee is running), refreshing the
disassembly flavor of the displayed memory region can replace the latest
region range with a previous one. The following steps reproduce the bug:
1. Prepare a debuggee that calls QThread::sleep(10) at some point.
2. Create two breakpoints: one before the sleep statement and another -
   after it.
3. Start debugging the debuggee with GDB and wait for the first
   breakpoint to be hit.
4. (either) Right-click an address greater than the address of the
   second breakpoint's statement in the disassembled region view of the
   Disassemble/Registers tool view, trigger the Change Address context
   menu action and press Enter.
4. (or) Ensure that the address of the second breakpoint's statement is
   absent from the memory region displayed when the first breakpoint is
   hit (i.e. that the breakpoint addresses are far enough apart).
5. Continue debugging; trigger an unchecked action in the Disassembly
   Flavor submenu of the disassembled region's context menu before the
   second breakpoint is hit.

Once the 10 seconds elapse, the debugger stops at the second breakpoint
and DisassembleWidget::update() queues a disassemble command that
changes the range of the displayed region to show the new current
address in it (1). Then the disassembly flavor change command queued
while the debuggee was running (sleeping) is executed and
DisassembleWidget::setDisassemblyFlavorHandler() queues a disassemble
command that refreshes the flavor of the currently displayed region (2).
Finally, the commands (1) and (2) are executed in order, and the second
command inadvertently overwrites the new region range applied by the
first one.

Introduce and use a new data member bool
DisassembleWidget::m_regionDisassemblyFlavorUpToDate to detect execution
of a disassemble command after the disassembly flavor change command and
cancel applying the obsolete result of the flavor-refreshing disassemble
command.