DisassembleWidget: update flavor actions (only) when needed
Currently the checked states of the actions in the Disassembly Flavor
submenu of the disassembled region's context menu are sometimes updated
when not needed and at other times not updated when needed. For example:
1. If the Disassemble/Registers tool view is visible when debugging
   starts, DisassembleWidget::slotActivate() is invoked before the debug
   session starts running, and therefore its call to
   DisassembleWidget::updateDisassemblyFlavor() returns early. Both
   flavor actions end up unchecked then - until the user hides and shows
   the tool view or the entire KDevelop window (due to spontaneous hide
   and show events).
2. The MI command `-gdb-show disassembly-flavor` is executed and the
   checked states of the flavor actions are updated each time the user
   hides and shows the tool view or the entire KDevelop window. Such
   work is usually redundant.
3. If the user triggers the Intel action in the Disassembly Flavor
   submenu, then starts another debug session while the previous one is
   running and the Disassemble/Registers tool view is visible, the
   actual disassembly flavor reverts to the default AT&T, but the Intel
   action remains checked.

Call DisassembleWidget::updateDisassemblyFlavor() from
DisassembleWidget::updateIfNeeded() if the tool view is visible to
ensure that the flavor actions are up-to-date whenever a disassembled
memory region is displayed.

Introduce yet another up-to-date bool data member in order to update the
flavor actions once per debug session.