DisassembleWindow: disable certain actions when selection is empty
Despite the SingleSelection selection mode, the selection of
DisassembleWindow can be empty for two reasons:
1) the current debug session has never reached the paused state, so the
   tree is still empty;
2) the user has deselected the single selected item by Ctrl-clicking it.

If the user triggers the context menu action Jump to Cursor while the
selection is empty, the following assertion failure occurs:
ASSERT failure in QList::at: "index out of range", file /usr/include/qt6/QtCore/qlist.h, line 479
The same applies to the context menu action Run to Cursor. The reason
for this failure is that DisassembleWidget::jumpToCursor() and
DisassembleWidget::runToCursor() assume that the list returned by
QTreeWidget::selectedItems() is nonempty and call at(0) on it.

Enable these two actions only if the selection is nonempty. Disable them
and the Change Address context menu action when DisassembleWindow's
selection changes to prevent the assertion failure or wrong behavior
because of newly empty or different selection. The selection can change
while the context menu is open if the current debug session transitions
from the active state to the paused state, which invokes
DisassembleWidget::update().