DebugController: store raw pointer, not QPointer to session
DebugController is the owner of debug sessions. A debug session may be
destroyed only when it transitions into the ended state - via the
deleteLater() call in DebugController::debuggerStateChanged(). If the
debug session that ends is the current session, debuggerStateChanged()
sets DebugController's current session pointer data member to null
before calling deleteLater() on it.

34adf3abb95d0df2865881d0aeb54889a2c763e3 introduced
QPointer<IDebugSession> DebugController::m_currentSession. Perhaps the
QPointer was needed back then due to a different ownership scheme. But
it is not useful anymore, because only the explicit clearing in
debuggerStateChanged() can set it to null. So switch to a raw pointer in
order to improve readability (no need to wonder how else might a debug
session be destroyed) and optimize.