RegistersManager: don't reset QScopedPointer to its current value
RegistersManager::architectureParsedSlot() resets QScopedPointer
m_registerController to a (usually) non-null value, then calls
setController(m_registerController.data()).
RegistersManager::setController() resets m_registerController with its
argument.

Resetting a QScopedPointer to its current value is safe because
QScopedPointer::reset() performs a test for self-reset and returns early
in this case. However, std::unique_ptr::reset() does not. So replacing
the QScopedPointer with a unique_ptr would cause a regression because of
the self-reset. Pass a new pointer to setController() directly instead
of reset() to avoid the treacherous self-reset.

Ensure that architectureParsedSlot() assigns a value to
m_currentArchitecture and prints debug output before calling
setController(), as before this commit.