RegistersManager: reparse architecture on current session change
When the user starts a new debug session while the previous one is still
running, a non-null new session replaces a non-null old session.
RegistersManager::setSession() reparses architecture only if a null
session becomes current - probably by mistake because if an architecture
change without restarting KDevelop is possible, then surely replacing a
debug session can bring such a change about.

The two data members of RegistersManager - m_currentArchitecture and
m_needToCheckArch - are redundant because m_registerController becomes
null when architecture needs to be reset and because the current
architecture is meaningfully used only in
RegistersManager::architectureParsedSlot() where it is available as the
argument. Remove these two data members and reparse architecture if
m_registerController is null instead.

Remove a no longer used enumerator Architecture::undefined. Replace the
label `default:` with the only remaining unsupported-architecture
enumerator Architecture::other in a switch over the enum.

Set m_registerController to null in RegistersManager::setSession() only
if the previous debug session is not null, because otherwise
m_registerController must already be null (set to null in the
constructor and when a non-null current debug session is replaced with
a null one). Assert that m_registerController is null to verify this
reasoning.

As m_registerController is always replaced when another session becomes
current, remove the unneeded function IRegisterController::setSession()
and make the pointer IRegisterController::m_debugSession const.