Require valid MIDebugSession for a register controller
IRegisterController::m_debugSession is a const pointer since a recent
commit titled "RegistersManager: reparse architecture on current session
change". Creating a useless register controller with a permanently null
debug session does not make sense. This also never happens because a
register controller can be created only in
RegistersManager::registerNamesHandler(), which handles a command to a
still current debug session.

Remove the default value of the debug session parameter of the
constructors of register controller classes. Assert that the debug
session is not null in registerNamesHandler() and in
IRegisterController(). Do not check m_debugSession for null. Document
the new requirement.

Also document a long-standing assumption that a debug session is not
destroyed before a register controller that stores a raw pointer to it.
RegistersManager::setSession() is invoked on current debug session
change and destroys an existing register controller (if any), and
thereby satisfies this second requirement.