Don't invoke tool views' handlers for previous debug session's commands
A recent commit titled "Disconnect from no longer current debug session"
introduced disconnection of debugger tool views from the signals and
slots of a no longer current debug session. But if a command added by a
tool view is being executed in the old session while the new session
replaces it, the handler of the command is still invoked once the old
debugger replies. The implementation of the tool views does not expect
such an invocation, so it can cause bugs (possibly since the recent
commits that modified related behavior of the tool views).

For example, a disassembled memory region from the previous session can
be misleadingly displayed during the new session, even though
DisassembleWidget::currentSessionChanged() clears it. Worse, if
RegistersManager::registerNamesHandler() is invoked for both the old and
the new session after RegistersManager::setSession(), the assertion
!m_registerController at the top of registerNamesHandler() would fail.

Introduce a new function template
MIDebugSession::addCommandWithCurrentSessionHandler() with a signature
identical to that of an existing function template overload of
MIDebugSession::addCommand(). The only difference is that using the new
function template prevents invoking the command handler if the command's
session is no longer the current session.

Use addCommandWithCurrentSessionHandler() in place of the addCommand()
overload with identical signature in the debugger tool views that can
interact with multiple debug sessions in order to prevent bugs and make
reasoning about the code easier. Also use the new function template in
place of a more specialized class ExpressionValueCommand in
MemoryView::changeMemoryRange(). This makes the MemoryView code slightly
more complex, but prevents reading a memory region in the new session
after computing its size in the old session.

Remove the no longer used class ExpressionValueCommand. The
implementation of the class is trivial, and using it is not
substantially more convenient compared to using addCommand() or
addCommandWithCurrentSessionHandler() directly. Furthermore, the class
encapsulates a command of the type DataEvaluateExpression that is issued
only in the single place in MemoryView (and now also in the added unit
test function). And besides, the implementation of the class is buggy:
it stores a QPointer to the handler object but never checks it for null.