MemoryView: compute sizeof(*pointer), not sizeof(pointer)
When the text of the Amount line edit of a Memory View widget is empty,
MemoryView::changeMemoryRange() computes sizeof(START) where START is
the text of the Start line edit. If START is not a pointer, the eventual
reading of a memory range fails and the memory view remains unchanged
(unfortunately no error message is reported in the UI, but the user can
enable the Show Internal Commands option in the context menu of the GDB
tool view and see the relevant debugger output, including error
messages, there).

So the result of the sizeof(START) computation and whether it succeeds
or fails matters only if START is a pointer. But the size of a pointer
is a constant for a given CPU architecture (usually 8 or 4 bytes). And
the Memory View widget displays the contents of a memory range that
starts at that pointer, not the pointer itself. So compute the size of
the value, i.e. of the dereferenced pointer, instead.