Introduce RegistersView::primaryActiveView() to optimize and document
The documentation for the new function explains RegistersView's
unobvious action-related code.

The function itself replaces and optimizes all uses of
activeViews().first(). Instead of splitting the current tab text into a
list of strings then calling QList::first() on the temporary list, which
detaches and copies it, primaryActiveView() efficiently computes and
returns the substring of the current tab text before the first view name
separator.

Extract a constant viewNameSeparator and a helper function
currentTabText() to avoid code duplication. Standardize on `group` as
the name of a local variable, to which primaryActiveView() is assigned.
Use more `const` and `auto`.