RegistersView: update registers of only the modified group
RegistersView::menuTriggered() changes the format or the mode of the
primary active view. But then calls RegistersView::updateRegisters(),
which issues MI commands to update registers of all active views. Pass
the primary group name to ModelsManager::updateRegisters() instead to
update registers of only the modified group and thus optimize.

RegistersView::updateRegisters() also calls
RegistersView::changeAvaliableActions(), which is not needed when only
the current format or mode changes. That's because changing the current
format or mode reorders the new current to the front of the list and
because action availability depends on the sets of formats and modes
irrespective of their orders.