Eliminate unneeded reset() signals of debugger session and plugin
MIDebugSession::stopDebugger() emits reset(), which invokes the emission
of MIDebuggerPlugin::reset(). Debugger output widgets connect their
slots clear() to the plugin's signal. Call clear() from the member
functions currentSessionChanged() of the debugger output widgets
instead. It is more logical to clear the output of the previous session
when it stops being current - along with disconnecting from its signals
and slots.

Now it is guaranteed and obvious that the output of the previous session
is cleared after disconnecting from its signals (so there is no way any
more output from it can arrive after the clearing) and before any output
of the next session can possibly arrive (before the signals of the newly
current session are connected to).