MIParser: tolerate an invalid MI message that starts with a number
Remove an assertion that a token number has not been parsed in an MI
message that does not conform to the specification:
https://sourceware.org/gdb/current/onlinedocs/gdb.html/GDB_002fMI-Output-Syntax.html
This eliminates an assertion failure while debugging a program with LLDB
when a breakpoint is placed in a shared library used by the program.

Add a test that triggers the assertion failure as follows:
QDEBUG : Debugger output (pid = 15638 ):  "1 location added to breakpoint 2"
QFATAL : ASSERT: "token == 0" in file kdevelop/plugins/debuggercommon/mi/miparser.cpp, line 86
FAIL!  : KDevMI::LLDB::LldbTest::testBreakpointInSharedLibrary() Received a fatal error.
   Loc: [kdevelop/plugins/debuggercommon/mi/miparser.cpp(86)]
But passes with the assertion removed.